+++ KrakenExchange/Functions/UserData/Get-KEOpenPositions.ps1
@@ -1,5 +1,39 @@
 function Get-KEOpenPositions {
     <#
+    .SYNOPSIS
+    Retrieves a list of open positions for the user's account on Kraken exchange.
+
+    .DESCRIPTION
+    The Get-KEOpenPositions function sends an authenticated request to the Kraken API to retrieve a list of open positions
+    for the user's account. This function requires an API key and secret to authenticate the request. If the API secret is
+    not provided, the function will prompt the user to enter their API key and secret interactively.
+
+    .PARAMETER ApiKey
+    The API key used to authenticate the request. This parameter is optional. If not provided, the function will attempt
+    to retrieve the API key from the 'KE_API_KEY' environment variable.
+
+    .PARAMETER ApiSecret
+    Encoded API secret used to authenticate the request. This parameter is optional. If not provided, the function will prompt
+    the user to enter their API key and secret interactively.
+
+    .PARAMETER txid
+    A comma delimited list of transaction IDs to restrict output to. This parameter is optional.
+
+    .PARAMETER docalcs
+    Whether or not to include profit/loss calculations. This parameter is optional and defaults to $false.
+
+    .PARAMETER consolidation
+    How to consolidate the positions. This parameter is optional and defaults to 'market'.
+
+    .EXAMPLE
+    PS C:> Get-KEOpenPositions -docalcs $true
+
+    Retrieves open positions and includes profit/loss calculation for each position.
+
+    .EXAMPLE
+    PS C:> Get-KEOpenPositions -txid 'O2TQZ6-KHJ6U-7SDXNW'
+
+    Retrieves open positions for a specific transaction ID.
 
     .LINK
     For more information, see the Kraken API documentation: